+++ app/backend/registration_data/identified_volunteers_at_event.py
@@ -160,3 +160,14 @@ def update_list_of_identified_volunteers_at_event(
         new_object=list_of_identified_volunteers_at_event,
         event_id=event.id,
     )
+
+def delete_volunteer_from_identified_data_and_return_rows_deleted(object_store: ObjectStore, event: Event, volunteer: Volunteer, areyousure=False):
+    if not areyousure:
+        return
+
+    list_of_identified_volunteers = get_list_of_identified_volunteers_at_event(object_store=object_store, event=event)
+    rows = list_of_identified_volunteers.list_of_identified_volunteers_with_volunteer_id(volunteer.id)
+    list_of_identified_volunteers.delete_all_rows_with_volunteer_id(volunteer.id)
+    update_list_of_identified_volunteers_at_event(object_store=object_store, list_of_identified_volunteers_at_event=list_of_identified_volunteers, event=event)
+
+    return len(rows)