+++ app/frontend/administration/admin_function_mapping.py
@@ -13,8 +13,9 @@ from app.frontend.administration.users.ENTRY_users import (
 from app.frontend.administration.data.data import display_form_data, post_form_data
 from app.frontend.administration.data.edit_delete_events import display_form_edit_delete_events, post_form_edit_delete_events
 from app.frontend.administration.data.merge_delete_cadets import display_form_merge_delete_cadets, post_form_merge_delete_cadets, display_form_merge_delete_individual_cadet, post_form_merge_delete_individual_cadet
-from app.frontend.administration.data.merge_delete_volunteers import display_form_merge_delete_volunteers, post_form_merge_delete_volunteers
+from app.frontend.administration.data.merge_delete_volunteers import display_form_merge_delete_volunteers, post_form_merge_delete_volunteers, display_form_merge_delete_individual_volunteer, post_form_merge_delete_individual_volunteer
 from app.frontend.administration.data.deleting_cadets_process import display_deleting_cadet_process, post_deleting_cadets_process
+from app.frontend.administration.data.deleting_volunteers_process import display_deleting_volunteer_process, post_deleting_volunteers_process
 
 admin_function_mapping = DisplayAndPostFormFunctionMaps.from_nested_dict_of_functions(
     NestedDictOfMappings(
@@ -27,7 +28,11 @@ admin_function_mapping = DisplayAndPostFormFunctionMaps.from_nested_dict_of_func
                             (display_deleting_cadet_process, post_deleting_cadets_process):0
                         }
                     },
-                    (display_form_merge_delete_volunteers, post_form_merge_delete_volunteers):0,
+                    (display_form_merge_delete_volunteers, post_form_merge_delete_volunteers):{
+                        (display_form_merge_delete_individual_volunteer, post_form_merge_delete_individual_volunteer): {
+                        (display_deleting_volunteer_process, post_deleting_volunteers_process):0
+                        }
+                    },
                     (display_form_edit_delete_events, post_form_edit_delete_events):0
                 },
             }